gcp-bookstore-data-project/bookstore-scrape-cloudfunc/requirements.txt
# cloudrun book-info-scrape
functions-framework==3.*
google-cloud-storage
requests
lxml